More namspace refactorings
